% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot3D.R
\name{plot3D}
\alias{plot3D}
\title{plot3D function to create a 3D plot of a ZooScatR shape}
\usage{
plot3D(para)
}
\arguments{
\item{para}{a list of parameters as used in ZooScatR}
}
\value{
a plotly plot object
}
\description{
Creates a 3D plotly plot of the prolate spheroid or input shape  
for ZooScatR, based on the parameters file.
}
\examples{

}
